Move files to directory
